requirements.txt
nltk~=3.9.1
spacy~=3.8.4
transformers~=4.48.2
rake-nltk~=1.0.6
yake
textblob~=0.19.0
vaderSentiment
pyspellchecker~=0.8.2
gramformer
language-tool-python
whoosh~=2.7.4
reportlab
python-docx
SpeechRecognition
pyttsx3
streamlit~=1.41.1
flask
flask-login
torch~=2.6.0
Wikipedia-API~=0.8.1
wikipedia~=1.4.0
GitPython~=3.1.44
gitdb~=4.0.12
numpy~=2.2.2
pillow~=11.1.0
pip~=25.0
attrs~=25.1.0
wheel~=0.41.2
protobuf~=5.29.3
filelock~=3.17.0
blis~=1.2.0
regex~=2024.11.6
click~=8.1.8
tqdm~=4.67.1
requests~=2.32.3
pytz~=2025.1
rich~=13.9.4
Pygments~=2.19.1
markdown-it-py~=3.0.0
toml~=0.10.2
colorama~=0.4.6
PyYAML~=6.0.2
typing_extensions~=4.12.2
cymem~=2.0.11
smmap~=5.0.2
mdurl~=0.1.2
charset-normalizer~=3.4.1
setuptools~=68.2.0
thinc~=8.3.4
srsly~=2.5.1
wasabi~=1.1.3
weasel~=0.4.1
typer~=0.15.1
catalogue~=2.0.10
Jinja2~=3.1.5
pydantic~=2.10.6
packaging~=24.2
langcodes~=3.5.0
functions~=0.7.0
preshed~=3.0.9
tornado~=6.4.2
sympy~=1.13.1
mpmath~=1.3.0
murmurhash~=1.0.12
confection~=0.1.5
tabulate~=0.9.0
urllib3~=2.3.0
fsspec~=2025.2.0
shellingham~=1.5.4
altair~=5.5.0
jsonschema~=4.23.0
narwhals~=1.25.0
pandas~=2.2.3
pyarrow~=19.0.0
referencing~=0.36.2
MarkupSafe~=3.0.2
joblib~=1.4.2
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
cloudpathlib~=0.20.0
certifi~=2025.1.31
tzdata~=2025.1
six~=1.17.0
networkx~=3.4.2
pydantic_core~=2.27.2
annotated-types~=0.7.0
idna~=3.10
tenacity~=9.0.0
watchdog~=6.0.0
cachetools~=5.5.1
blinker~=1.9.0
jsonschema-specifications~=2024.10.1
smart-open~=7.1.0
wrapt~=1.17.2
tokenizers~=0.21.0
safetensors~=0.5.2
spacy-legacy~=3.0.12
huggingface-hub~=0.28.1
beautifulsoup4~=4.13.1
marisa-trie~=1.2.1
language_data~=1.3.0
spacy-loggers~=1.0.5